+++ src/elements/conditions/contacts/EmailConditionRule.php
@@ -1,0 +1,54 @@
+<?php
+/**
+ * @copyright Copyright (c) PutYourLightsOn
+ */
+
+namespace putyourlightson\campaign\elements\conditions\contacts;
+
+use Craft;
+use craft\base\conditions\BaseTextConditionRule;
+use craft\base\ElementInterface;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\elements\db\ElementQueryInterface;
+use putyourlightson\campaign\elements\ContactElement;
+use putyourlightson\campaign\elements\db\ContactElementQuery;
+
+/**
+ * @since 2.0.0
+ */
+class EmailConditionRule extends BaseTextConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('campaign', 'Email');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['email'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(ElementQueryInterface $query): void
+    {
+        /** @var ContactElementQuery $query */
+        $query->email($this->paramValue());
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        /** @var ContactElement $element */
+        return $this->matchValue($element->email);
+    }
+}