+++ src/elements/db/SegmentElementQuery.php
@@ -41,6 +41,7 @@ class SegmentElementQuery extends ElementQuery
 
         $this->query->select([
             'campaign_segments.segmentType',
+            'campaign_segments.contactCondition',
             'campaign_segments.conditions',
             'campaign_segments.template',
         ]);